backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.26
langgraph==0.5.3
voyageai==0.3.3
pinecone==6.0.0
pytest==8.4.1
pytest-env==1.1.5
ragas==0.3.0
langsmith==0.4.8
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
langchain-google-genai==2.0.8
langchain-pinecone==0.2.9  # New
python-dotenv==1.1.1
pandas==2.2.3
requests==2.32.3